Adds Vertex representation as a dungeon room (a rectangle in space). Creates a dungeon with random rooms and connect them using delaunay triangulation. Adds class to draw dungeon as a JFrame.

+++ src/graph/Edge.java
@@ -9,23 +9,23 @@ public class Edge {
     }
 
     private Vertex destination;
-    private int weight;
+    private float weight;
 
     public Edge(Vertex destination) {
         setDestination(destination);
         setWeight(1);
     }
 
-    public Edge(Vertex destination, int value) {
+    public Edge(Vertex destination, float value) {
         setDestination(destination);
         setWeight(value);
     }
 
-    public int getWeight() {
+    public float getWeight() {
         return weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(float weight) {
         this.weight = weight;
     }
 }
